api/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
psycopg==3.3.2
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.4
